+++ lib/feature/presentation/widgets/splash_screen.dart
@@ -3,7 +3,7 @@ import 'dart:async';
 import 'package:animated_text_kit/animated_text_kit.dart';
 import 'package:flutter/material.dart';
 import 'package:lottie/lottie.dart';
-import 'package:planner/feature/presentation/screens/home_screen.dart';
+import 'package:planner/feature/presentation/navigation/main_navigation.dart';
 
 class SplashScreen extends StatefulWidget {
   const SplashScreen({super.key});
@@ -18,10 +18,9 @@ class _SplashScreenState extends State<SplashScreen> {
     super.initState();
     Timer(
       const Duration(seconds: 3),
-      () => Navigator.of(context).push(
-        MaterialPageRoute(
-          builder: (context) => const HomeScreen(),
-        ),
+      () => Navigator.of(context).pushNamed(
+        MainNavigationRouteNames.homeScreen,
+        arguments: true,
       ),
     );
   }
