+++ lib/main.dart
@@ -1,22 +1,8 @@
 import 'package:flutter/material.dart';
-import 'package:planner/feature/presentation/widgets/splash_screen.dart';
+import 'package:planner/feature/presentation/screens/app/my_app.dart';
 
 void main() {
   runApp(const MyApp());
 }
 
-class MyApp extends StatelessWidget {
-  const MyApp({super.key});
-  @override
-  Widget build(BuildContext context) {
-    return MaterialApp(
-      title: 'Flutter Demo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: Colors.deepPurple),
-        useMaterial3: true,
-      ),
-      debugShowCheckedModeBanner: false,
-      home: const SplashScreen(),
-    );
-  }
-}
+
